requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
click==8.3.1
colorama==0.4.6
fastapi==0.123.8
graphql-core==3.2.7
h11==0.16.0
idna==3.11
joblib==1.5.2
lia-web==0.2.3
libcst==1.8.6
markdown-it-py==4.0.0
mdurl==0.1.2
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
rich==14.2.0
scikit-learn==1.7.2
scipy==1.16.3
shellingham==1.5.4
six==1.17.0
starlette==0.50.0
strawberry-graphql==0.287.2
threadpoolctl==3.6.0
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.38.0
websockets==15.0.1
